[chore] int -> long으로 변경

+++ src/main/java/com/wayble/server/explore/dto/common/WaybleZoneInfoResponseDto.java
@@ -14,7 +14,7 @@ public record WaybleZoneInfoResponseDto(
         Double latitude,
         Double longitude,
         Double averageRating,
-        int reviewCount,
+        long reviewCount,
         FacilityResponseDto facility
 ) {
     public static WaybleZoneInfoResponseDto from(WaybleZoneDocument document) {
